The app is functional, but....
The app functions (not good), it spits out words but some files give more words than it should give so I need to look into the word cleaning but if anything the app is presentable.
Also moved some files around so I can test the files that give problems and Identify if it's an issue with the code or with the files.